Make VirtualCanvasNode sealed
