--- app/src/main/java/io/github/karino2/kotlitex/renderer/node/VerticalListRow.kt
@@ -1,4 +1,0 @@
-package io.github.karino2.kotlitex.renderer.node
-
-class VerticalListRow(klasses: Set<String>) : VirtualContainerNode<VirtualCanvasNode>(klasses) {
-}
